+++ app/(auth)/login/page.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import "../globals.css";
+import "../../globals.css";
 import { Button } from "@/components/ui/button";
 
-export default function Signup() {
+export default function Login() {
   return (
     <div className="w-full flex px-[160px] py-20">
       <div className="m-0 w-full flex flex-col h-[695px] justify-center items-center">
@@ -21,15 +21,15 @@ export default function Signup() {
           name="password"
           id="password"
           placeholder="パスワード"
-          autoComplete="new-password"
+          autoComplete="current-password"
           className="w-[448px] bg-[#F5F2F2] mx-[12px] my-[16px] p-[16px] rounded-[12px] placeholder:text-[16px] placeholder:text-[#827066]"
         />
-        <Button variant="loginSignup">登録</Button>
+        <Button variant="loginSignup">ログイン</Button>
         <a
-          href="/login"
+          href="/signup"
           className="cursor-pointer pt-[4px] pb-[12px] px-[16px] text-[14px] text-[#827066]"
         >
-          既にアカウントをお持ちですか？ ログイン
+          アカウントをお持ちでないですか？ 登録
         </a>
       </div>
     </div>
