+++ app/(auth)/signup/page.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import "../globals.css";
+import "../../globals.css";
 import { Button } from "@/components/ui/button";
 
-export default function Login() {
+export default function Signup() {
   return (
     <div className="w-full flex px-[160px] py-20">
       <div className="m-0 w-full flex flex-col h-[695px] justify-center items-center">
@@ -21,15 +21,15 @@ export default function Login() {
           name="password"
           id="password"
           placeholder="パスワード"
-          autoComplete="current-password"
+          autoComplete="new-password"
           className="w-[448px] bg-[#F5F2F2] mx-[12px] my-[16px] p-[16px] rounded-[12px] placeholder:text-[16px] placeholder:text-[#827066]"
         />
-        <Button variant="loginSignup">ログイン</Button>
+        <Button variant="loginSignup">登録</Button>
         <a
-          href="/signup"
+          href="/login"
           className="cursor-pointer pt-[4px] pb-[12px] px-[16px] text-[14px] text-[#827066]"
         >
-          アカウントをお持ちでないですか？ 登録
+          既にアカウントをお持ちですか？ ログイン
         </a>
       </div>
     </div>
